参考帧列表初始化

+++ ParseSPS.h
@@ -43,8 +43,8 @@ public:
 	bool qpprime_y_zero_transform_bypass_flag;  // 0 u(1)
 	bool seq_scaling_matrix_present_flag;  // 0 u(1)
 	bool  seq_scaling_list_present_flag[12]; //seq_scaling_list_present_flag[ i ] 0 u(1)
-	uint32_t log2_max_frame_num_minus4;  // 0 ue(v)
-	uint32_t MaxFrameNum;
+	uint8_t log2_max_frame_num_minus4;  // 0-12
+	uint16_t MaxFrameNum;
 	uint32_t pic_order_cnt_type;  // 0 ue(v)
 	uint32_t log2_max_pic_order_cnt_lsb_minus4;  // 0 ue(v)
 	uint32_t  MaxPicOrderCntLsb;
